+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added
 - GC and exception markers to the chart
+- Automatically attach charts to a starting process 
 
 ## [2023.2.0] - 2023-08-07